Updates Percent() as per discussion and also renames player-card-line to something more generic for easier moving later on.
